Add message threading support to slackbot
Update num_turns() to return and int rather than a float

+++ chatai.py
@@ -75,7 +75,7 @@ class Conversation:
 
     def num_turns(self):
         """Return the number of turns in the conversation"""
-        return (len(self.turns)-1) / 2
+        return int((len(self.turns)-1) / 2)
 
     def __str__(self):
         conversation_str = ""
